make unit price decimal in OrderRespnse

+++ tickethub/Dtos/Order/OrderResponse.cs
@@ -4,7 +4,7 @@ public record OrderResponse(
     Guid Id,
     string Email,
     int Qty,
-    int UnitPrice,
+    decimal UnitPrice,
     decimal TotalPrice,
     int ConcertId
 );
